// src/api/type/create-ledger-form.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Create a new Ledger. Use this schema to create a ledger definition and POST it back to the CREATE_URI endpoint.
 */
export interface CreateLedgerForm {
  /**
   * The name of the ledger.
   */
  name: string
  /**
   * A description of the ledger's purpose.
   */
  description: string
}