// src/api/type/download-ledger-lookup.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Download a ledger's events, either entirely or after an event ID. Use this schema to create a download command and POST it back to DOWNLOAD_URI endpoint. To download all the events, send an empty JSON object in the request.
 */
export interface DownloadLedgerLookup {
  /**
   * Download events that exist after this event ID.
   */
  after?: string
  /**
   * Limit the number of returned events to this value.
   */
  limit?: number
}