// src/api/type/reset-ledger-form.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Form to reset a ledger's events, either entirely or after an event ID. Use this schema to create a reset command and POST it back to RESET_URI endpoint.
 */
export interface ResetLedgerForm {
  /**
   * Remove all events that occur after this event ID.
   */
  after?: string
}